split initial load table into 2
